requirements.txt
polyscope
sympy
numpy
libigl
scipy
